sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "layer5"
version = "0.1.0"
description = "Outcome-ranked decision intelligence for AI agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = [
    "ai", "agents", "llm", "langchain", "crewai",
    "autogen", "decision-intelligence", "mlops",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2.0"]
llamaindex = ["llama-index-core>=0.10.0"]
crewai = ["crewai>=0.28.0"]
autogen = ["pyautogen>=0.2.0"]
openai = ["openai>=1.0.0"]
all = [
    "langchain-core>=0.2.0",
    "llama-index-core>=0.10.0",
    "crewai>=0.28.0",
    "pyautogen>=0.2.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "langchain-core>=0.2.0",
    "llama-index-core>=0.10.0",
    "crewai>=0.28.0",
    "openai>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py39"
